requirements.txt
streamlit
opencv-python
numpy
imagehash
Pillow
reportlab
